refactor: Phase 3 - modularize App.tsx into focused components
Break down the monolithic 779-line App.tsx into focused, maintainable components:

**New Components:**
- AppStateProvider.tsx (249 lines) - Centralized state management through React Context
- ModalManager.tsx (99 lines) - Manages modal overlays with z-index priority
- StatusBarManager.tsx (92 lines) - Unified status bars and interface warnings
- ContentRouter.tsx (135 lines) - Routes between different view modes
- KeyboardManager.tsx (105 lines) - Centralized keyboard input handling

**Improvements:**
- Refactored App.tsx from 779 lines to 184 lines using modular architecture
- Applied single responsibility principle to each component
- Maintained all existing functionality and keyboard navigation
- Preserved clean architecture patterns established in Phase 1 & 2
- All tests pass (793 passed, 12 skipped)

**Architecture Benefits:**
- Clear separation of concerns between state, UI, and interaction layers
- Improved maintainability and testability
- Follows established patterns from previous phases
- Ready for future feature development
